+++ refactor/executor/executor_plugins/local.py
@@ -37,8 +37,8 @@ from covalent._shared_files.util_classes import DispatchInfo
 from covalent._workflow.transport import TransportableObject
 from covalent.executor import BaseExecutor
 
-# The plugin class name must be given by the executor_plugin_name attribute:
-executor_plugin_name = "LocalExecutor"
+# The plugin class name must be given by the EXECUTOR_PLUGIN_NAME attribute:
+EXECUTOR_PLUGIN_NAME = "LocalExecutor"
 
 app_log = logger.app_log
 log_stack_info = logger.log_stack_info