+++ CHANGELOG.md
@@ -18,6 +18,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - In covalent/executor/__init__.py, `from importlib import metadata` is used instead of `importlib.metadata`.
 - Electron.get_op_function.rename now uses the correct separator string when renaming a function.
 
+## [0.56.1] - 2022-04-05
+
+### Fixed
+
+- Fixed various module import errors in the containers for the microservices.
+
+### Tests
+
+- Added tests for post-refactor covalent cli commands: start, stop, restart, status, and logs
+
+## [0.56.0] - 2022-04-05
+
+### Changed
+
+- Changed global variable executor_plugin_name to EXECUTOR_PLUGIN_NAME in executors to conform with PEP8.
+
+## [0.55.0] - 2022-04-04
+
+### Changed
+
+- Changed supervisord http server's default to listen on all interfaces, so that covalent can run on any computer in a trusted LAN (without firewalls/auth).
+
+## [0.54.0] - 2022-04-04
+
+### Added
+
+- Draw workflow draft API to ui_backend service
+
+
 ## [0.53.0] - 2022-04-04
 
 ### Added